hotel_booking_system/requirements.txt
pandas
matplotlib
seaborn
sentence-transformers
faiss-cpu
fastapi
uvicorn
pydantic
numpy